+++ packages/outline/src/__tests__/unit/OutlineEditor.test.js
@@ -21,7 +21,7 @@ import {
 } from 'outline';
 import {createParagraphNode, ParagraphNode} from 'outline/ParagraphNode';
 import useOutlineRichText from 'outline-react/useOutlineRichText';
-import {getNodeByKey} from '../../core/OutlineProcess';
+import {getNodeByKey} from '../../core/OutlineUtils';
 import {createTestBlockNode} from '../utils';
 
 describe('OutlineEditor tests', () => {
